Proposes new link
The previous link didn't work and the domain actually seems to not exist anymore.
The book that the broken link seems to refer to is [this one](https://theswissbay.ch/pdf/Gentoomen%20Library/Databases/Molina%2CUllman%20-%20Database%20Systems%20The%20Complete%20Book.pdf) I guess. But it is too long and I can't find a quick way to point to chapter 3 only. 
Instead, I propose [this link](http://database.guide/the-3-types-of-relationships-in-database-design/), which seems to be a quick and comprehensive read about the different types of database relationships.